janus_new/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
supabase-py
pydantic-settings
python-jose[cryptography] # Added for password generation
